use failure variant for failures
Previously, the KV_RO implementation in mirage-fs-unix would claim that any failure of the
underlying filesystem should be interpreted as an unknown key.  Stop
doing that, and instead report the error message via the "Failure"
variant type.

This pattern should be replaced entirely with result types in the near
future.